reduce code complexity
